requirements.txt
discord.py==2.3.2
anthropic==0.40.0
beautifulsoup4==4.12.3
requests==2.32.3
apscheduler==3.10.4
python-dotenv==1.0.1
aiohttp==3.11.11
